+++ dev/cpp/emscripten.cpp
@@ -509,6 +509,7 @@ extern "C" {
     void loadTrainingData (int instanceIndex, float *buf, int total, int size, int dimension) {
         Network* net = Network::getInstance(instanceIndex);
         net->trainingData.clear();
+        net->collectErrors = false;
 
         std::tuple<std::vector<double>, std::vector<double> > epoch;
 
@@ -527,6 +528,65 @@ extern "C" {
                 std::get<1>(epoch).push_back((double)buf[i]);
             }
         }
+    }
+
+    EMSCRIPTEN_KEEPALIVE
+    double* get_collected_training (int instanceIndex) {
+
+        Network* net = Network::getInstance(instanceIndex);
+
+        int errorsCount = net->collectedTrainingErrors.size();
+        double errors[errorsCount+1];
+        errors[0] = errorsCount;
+
+
+        for (int i=1; i<=errorsCount; i++) {
+            errors[i] = net->collectedTrainingErrors[i];
+        }
+
+        auto ptr = &errors[0];
+        return ptr;
+    }
+
+    EMSCRIPTEN_KEEPALIVE
+    double* get_collected_test (int instanceIndex) {
+
+        Network* net = Network::getInstance(instanceIndex);
+
+        int errorsCount = net->collectedTestErrors.size();
+        double errors[errorsCount+1];
+        errors[0] = errorsCount;
+
+
+        for (int i=1; i<=errorsCount; i++) {
+            errors[i] = net->collectedTestErrors[i];
+        }
+
+        auto ptr = &errors[0];
+        return ptr;
+    }
+
+    EMSCRIPTEN_KEEPALIVE
+    double* get_collected_validation (int instanceIndex) {
+
+        Network* net = Network::getInstance(instanceIndex);
+
+        int errorsCount = net->collectedValidationErrors.size();
+        double errors[errorsCount+1];
+        errors[0] = errorsCount;
+
+
+        for (int i=1; i<=errorsCount; i++) {
+            errors[i] = net->collectedValidationErrors[i];
+        }
+
+        auto ptr = &errors[0];
+        return ptr;
+    }
+
+    EMSCRIPTEN_KEEPALIVE
+    void collectErrors (int instanceIndex) {
+        Network::getInstance(instanceIndex)->collectErrors = true;
     }
 
     EMSCRIPTEN_KEEPALIVE
@@ -569,6 +629,7 @@ extern "C" {
     void loadTestingData (int instanceIndex, float *buf, int total, int size, int dimension) {
         Network* net = Network::getInstance(instanceIndex);
         net->testData.clear();
+        net->collectErrors = false;
         std::tuple<std::vector<double>, std::vector<double> > epoch;
 
         // Push test data to memory