+++ dev/js-WebAssembly/Network.js
@@ -229,6 +229,11 @@ class Network {
         NetUtil.defineProperty(this, "earlyStoppingPatience", ["number"], [this.netInstance])
         NetUtil.defineProperty(this, "earlyStoppingPercent", ["number"], [this.netInstance])
 
+        this.collectedErrors = {}
+        NetUtil.defineArrayProperty(this.collectedErrors, "training", ["number"], [this.netInstance], "auto", {pre: "collected_"})
+        NetUtil.defineArrayProperty(this.collectedErrors, "test", ["number"], [this.netInstance], "auto", {pre: "collected_"})
+        NetUtil.defineArrayProperty(this.collectedErrors, "validation", ["number"], [this.netInstance], "auto", {pre: "collected_"})
+
         if (layers.length) {
 
             this.state = "constructed"
@@ -342,7 +347,7 @@ class Network {
         })
     }
 
-    train (data, {epochs=1, callback, callbackInterval=1, miniBatchSize=1, log=true, shuffle=false, validation}={}) {
+    train (data, {epochs=1, callback, callbackInterval=1, collectErrors, miniBatchSize=1, log=true, shuffle=false, validation}={}) {
 
         miniBatchSize = typeof miniBatchSize=="boolean" && miniBatchSize ? data[0].expected.length : miniBatchSize
         this.Module.ccall("set_miniBatchSize", null, ["number", "number"], [this.netInstance, miniBatchSize])
@@ -384,6 +389,10 @@ class Network {
 
             if (shuffle) {
                 this.Module.ccall("shuffleTrainingData", null, ["number"], [this.netInstance])
+            }
+
+            if (collectErrors) {
+                this.Module.ccall("collectErrors", null, ["number"], [this.netInstance])
             }
 
             let validationBuf
@@ -575,7 +584,7 @@ class Network {
         }
     }
 
-    test (data, {log=true, callback}={}) {
+    test (data, {log=true, collectErrors, callback}={}) {
         return new Promise((resolve, reject) => {
 
             if (data === undefined || data === null) {
@@ -599,6 +608,10 @@ class Network {
 
             this.Module.ccall("loadTestingData", "number", ["number", "number", "number", "number", "number"],
                                             [this.netInstance, buf, itemsCount, itemSize, dimension])
+
+            if (collectErrors) {
+                this.Module.ccall("collectErrors", null, ["number"], [this.netInstance])
+            }
 
             if (callback) {