+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added confusion matrix
 - Made it possible to pass volume (3D array) input data
 - Added callback interval config to .train()
+- Added collectErrors config to .train() and .test()
 
 #### InputLayer
 - Added InputLayer class